Rename jwt_helper.py to jwt_helpers.py
